requirements.txt
scrapy==2.5.0
dateparser==1.1.1
python-dateutil==2.8.1
beautifulsoup4==4.9.3
requests==2.13.0
